refactor(play): extract InducementsPhaseUI + normalizeState (S26.0a)
Premiere tranche du refactor S26.0 (PREREQUIS sprint S26) :
play/[id]/page.tsx 1666 -> 1559 lignes.

- Extrait `InducementsPhaseUI` (~98 lignes) dans
  apps/web/app/play/[id]/components/InducementsPhaseUI.tsx avec
  interface props nommee.
- Extrait `normalizeState` (utilise 16 fois dans page.tsx + le
  composant) dans apps/web/app/play/[id]/utils/normalize-state.ts
  pour permettre l'import depuis le composant extrait.
- page.tsx perd ~107 lignes + 4 imports inutilises (InducementSelector,
  INDUCEMENT_CATALOGUE, InducementSelection, InducementDefinition).

Slices restantes pour S26.0 : extraire SetupPhaseUI, MatchEndUI,
BlockChoiceFlow, ThrowTeamMateFlow, puis types Move unions pour
eliminer les `as any` (13 occurrences restantes).

Fix collateral : durcit les types apiRequest dans
apps/web/app/me/teams/[id]/edit/page.tsx (positionsData
unknown[] -> AvailablePosition[], result.advancement unknown ->
{skillSlug:string}). Errors TS pre-existantes liees aux
feature-flags ne sont PAS dans ce diff.
